Registering wrapupcode resource with framework instead of plugin-sdk

+++ genesyscloud/routing_wrapupcode_v2/routing_wrapupcode_v2.go
@@ -9,8 +9,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/mypurecloud/platform-client-sdk-go/v157/platformclientv2"
-	"github.com/mypurecloud/terraform-provider-genesyscloud/genesyscloud/provider"
 	"log"
+	"os"
+	"strconv"
 )
 
 var (
@@ -61,7 +62,6 @@ func (r *WrapupCodeResource) Metadata(_ context.Context, req resource.MetadataRe
 }
 
 func (r *WrapupCodeResource) Create(ctx context.Context, request resource.CreateRequest, response *resource.CreateResponse) {
-	log.Println("Calling create")
 	if r.clientConfig == nil {
 		response.Diagnostics.AddError(
 			"Uninitialised configuration object",
@@ -79,16 +79,9 @@ func (r *WrapupCodeResource) Create(ctx context.Context, request resource.Create
 		return
 	}
 
-	wrapupCodeReqBody := platformclientv2.Wrapupcoderequest{
-		Name:        data.Name.ValueStringPointer(),
-		Description: data.Description.ValueStringPointer(),
-	}
+	wrapupCodeReqBody := buildWrapupCodeRequestBody(data)
 
-	if !data.DivisionId.IsNull() {
-		wrapupCodeReqBody.Division.Id = data.DivisionId.ValueStringPointer()
-	}
-
-	log.Printf("Creating wrap up code '%s'", data.Name.ValueString())
+	frameworkLog("Creating wrap up code " + strconv.Quote(data.Name.ValueString()))
 	respObj, apiResponse, err := api.PostRoutingWrapupcodes(wrapupCodeReqBody)
 	if err != nil {
 		var apiResponseDetails = "No API response data returned"
@@ -98,31 +91,14 @@ func (r *WrapupCodeResource) Create(ctx context.Context, request resource.Create
 		response.Diagnostics.AddError("failed to create wrap up code", fmt.Sprintf("API Response: %s. Error: %s.", apiResponseDetails, err.Error()))
 		return
 	}
-	log.Printf("Successfully created wrap up code '%s'. ID: '%s'", *respObj.Name, *respObj.Id)
+	frameworkLog(fmt.Sprintf("Successfully created wrap up code '%s'. ID: '%s'", *respObj.Name, *respObj.Id))
 
-	data.Id = types.StringValue(*respObj.Id)
-	data.Name = types.StringValue(*respObj.Name)
-	if respObj.Description != nil {
-		data.Description = types.StringValue(*respObj.Description)
-	}
-	if respObj.Division != nil && respObj.Division.Id != nil {
-		data.DivisionId = types.StringValue(*respObj.Division.Id)
-	}
+	flattenWrapupCodeResponse(&data, *respObj)
 
 	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
 }
 
 func (r *WrapupCodeResource) Read(ctx context.Context, request resource.ReadRequest, response *resource.ReadResponse) {
-	log.Println("Calling read")
-}
-
-func (r *WrapupCodeResource) Update(ctx context.Context, request resource.UpdateRequest, response *resource.UpdateResponse) {
-	log.Println("Calling update")
-}
-
-func (r *WrapupCodeResource) Delete(ctx context.Context, request resource.DeleteRequest, response *resource.DeleteResponse) {
-	log.Println("Calling delete")
-
 	if r.clientConfig == nil {
 		response.Diagnostics.AddError(
 			"Uninitialised configuration object",
@@ -139,7 +115,77 @@ func (r *WrapupCodeResource) Delete(ctx context.Context, request resource.Delete
 		return
 	}
 
-	log.Printf("Deleting wrap up code '%s'", data.Name.ValueString())
+	respBody, resp, err := api.GetRoutingWrapupcode(data.Id.ValueString())
+	if err != nil {
+		var apiResponseDetails = "No API response data returned"
+		if resp != nil {
+			apiResponseDetails = resp.String()
+		}
+		response.Diagnostics.AddError(
+			"failed to read wrap up code "+strconv.Quote(data.Id.ValueString()),
+			fmt.Sprintf("API Response: %s. Error: %s.", apiResponseDetails, err.Error()),
+		)
+		return
+	}
+
+	flattenWrapupCodeResponse(&data, *respBody)
+
+	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
+}
+
+func (r *WrapupCodeResource) Update(ctx context.Context, request resource.UpdateRequest, response *resource.UpdateResponse) {
+	if r.clientConfig == nil {
+		response.Diagnostics.AddError(
+			"Uninitialised configuration object",
+			"Expected r.clientConfig to not be nil",
+		)
+		return
+	}
+
+	api := platformclientv2.NewRoutingApiWithConfig(r.clientConfig)
+
+	var data WrapupCodeResourceApiModel
+	response.Diagnostics.Append(request.Plan.Get(ctx, &data)...)
+	if response.Diagnostics.HasError() {
+		return
+	}
+
+	reqBody := buildWrapupCodeRequestBody(data)
+
+	frameworkLog("Updating wrap up code " + strconv.Quote(data.Name.ValueString()))
+	_, resp, err := api.PutRoutingWrapupcode(data.Id.ValueString(), reqBody)
+	if err != nil {
+		var apiResponseDetails = "No API response data returned"
+		if resp != nil {
+			apiResponseDetails = resp.String()
+		}
+		response.Diagnostics.AddError("failed to update wrap up code", fmt.Sprintf("API Response: %s. Error: %s.", apiResponseDetails, err.Error()))
+		return
+	}
+
+	frameworkLog("Successfully updated wrap up code " + strconv.Quote(data.Name.ValueString()))
+
+	response.Diagnostics.Append(response.State.Set(ctx, &data)...)
+}
+
+func (r *WrapupCodeResource) Delete(ctx context.Context, request resource.DeleteRequest, response *resource.DeleteResponse) {
+	if r.clientConfig == nil {
+		response.Diagnostics.AddError(
+			"Uninitialised configuration object",
+			"Expected r.clientConfig to not be nil",
+		)
+		return
+	}
+
+	api := platformclientv2.NewRoutingApiWithConfig(r.clientConfig)
+
+	var data WrapupCodeResourceApiModel
+	response.Diagnostics.Append(request.State.Get(ctx, &data)...)
+	if response.Diagnostics.HasError() {
+		return
+	}
+
+	frameworkLog("Deleting wrap up code " + strconv.Quote(data.Name.ValueString()))
 	apiResponse, err := api.DeleteRoutingWrapupcode(data.Id.ValueString())
 	if err != nil {
 		apiResponseDetails := "No API response details returned."
@@ -149,7 +195,8 @@ func (r *WrapupCodeResource) Delete(ctx context.Context, request resource.Delete
 		response.Diagnostics.AddError("Failed to delete wrap up code", fmt.Sprintf("API Response: %s. Error: %s", apiResponseDetails, err.Error()))
 		return
 	}
-	log.Printf("Successfully deleted wrap up code '%s'", data.Name.ValueString())
+
+	frameworkLog("Successfully deleted wrap up code " + strconv.Quote(data.Name.ValueString()))
 }
 
 func (r *WrapupCodeResource) Configure(ctx context.Context, req resource.ConfigureRequest, resp *resource.ConfigureResponse) {
@@ -157,18 +204,28 @@ func (r *WrapupCodeResource) Configure(ctx context.Context, req resource.Configu
 		return
 	}
 
-	client, ok := req.ProviderData.(provider.GenesysCloudProvider)
-	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected *GenesysCloudProvider, got: %T. Please report this issue to the provider developers.", req.ProviderData),
-		)
-		return
-	}
+	// TODO: Figure out cyclic issue between this package and 'provider' package
 
-	clientConfig, err := provider.AcquireSdkClient(ctx, client)
+	//
+	//client, ok := req.ProviderData.(provider.GenesysCloudProvider)
+	//if !ok {
+	//	resp.Diagnostics.AddError(
+	//		"Unexpected Resource Configure Type",
+	//		fmt.Sprintf("Expected *GenesysCloudProvider, got: %T. Please report this issue to the provider developers.", req.ProviderData),
+	//	)
+	//	return
+	//}
+	//
+	//clientConfig, err := provider.AcquireSdkClient(ctx, client)
+	//if err != nil {
+	//	resp.Diagnostics.AddError("Failed to acquire sdk client from pool", err.Error())
+	//	return
+	//}
+
+	clientConfig := platformclientv2.GetDefaultConfiguration()
+
+	err := clientConfig.AuthorizeClientCredentials(os.Getenv("GENESYSCLOUD_OAUTHCLIENT_ID"), os.Getenv("GENESYSCLOUD_OAUTHCLIENT_SECRET"))
 	if err != nil {
-		resp.Diagnostics.AddError("Failed to acquire sdk client from pool", err.Error())
 		return
 	}
 
@@ -177,4 +234,30 @@ func (r *WrapupCodeResource) Configure(ctx context.Context, req resource.Configu
 
 func NewWrapupCodeResource() resource.Resource {
 	return &WrapupCodeResource{}
+}
+
+func buildWrapupCodeRequestBody(data WrapupCodeResourceApiModel) platformclientv2.Wrapupcoderequest {
+	wrapupCodeReqBody := platformclientv2.Wrapupcoderequest{
+		Name:        data.Name.ValueStringPointer(),
+		Description: data.Description.ValueStringPointer(),
+	}
+	if !data.DivisionId.IsNull() {
+		wrapupCodeReqBody.Division.Id = data.DivisionId.ValueStringPointer()
+	}
+	return wrapupCodeReqBody
+}
+
+func flattenWrapupCodeResponse(data *WrapupCodeResourceApiModel, respBody platformclientv2.Wrapupcode) {
+	data.Id = types.StringValue(*respBody.Id)
+	data.Name = types.StringValue(*respBody.Name)
+	if respBody.Description != nil {
+		data.Description = types.StringValue(*respBody.Description)
+	}
+	if respBody.Division != nil && respBody.Division.Id != nil {
+		data.DivisionId = types.StringValue(*respBody.Division.Id)
+	}
+}
+
+func frameworkLog(s string) {
+	log.Println("(Framework) ", s)
 }
